fifo/gm20b: kick channel during cleanup
GM20B requires a channel kick to be performed during gpfifo cleanup, or
the FIFO will attempt to fetch memory from the previous context as a
channel is recycled. A previous commit attempted to do this for all
Kepler GPUs, but due to bug reports that pinned it down it has been
reverted. The present commit limits its scope to GM20B only.

The only effective change of this patch is to add a call to
gk104_fifo_gpfifo_kick() in gpfifo_fini for GM20B, but doing so requires
to export quite a few extra functions, hence its non-trivial length.

Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>